Move counter tests to TurnCounterTest
